artefact_scanner/requirements.txt
boto3==1.34.84
peewee==3.17.1
peewee-plus==1.3.0
psycopg2==2.9.9
uuid==1.30
pillow==9.0.1
